+++ jsync_rs/crates/jsync/src/lib.rs
@@ -1,4 +1,6 @@
-//! Jsync message consumer.
+//! Jsync message consumer and producer.
+
+mod producer;
 
 mod consumer;
 mod error;
@@ -6,3 +8,4 @@ mod value;
 
 pub use consumer::Consumer;
 pub use error::{JsyncError, JsyncErrorKind};
+pub use producer::Producer;